+++ main.go
@@ -18,24 +18,6 @@ func main() {
 	SetupMultiplexer()
 }
 
-func SetupMultiplexer() {
-	http.HandleFunc("/api/v1/suggestions/materials", GetMaterialSuggestions)
-
-	if err := http.ListenAndServe("localhost:9000", nil); err != nil {
-		log.Fatalln("There was an error starting server: ", err)
-	}
-}
-
-func GetMaterialSuggestions(res http.ResponseWriter, req *http.Request) {
-	cards := []Card{
-		{CardName: "Elemental HERO Avian"},
-		{CardName: "Elemental HERO Burstinatrix"},
-	}
-
-	res.Header().Add("Content-Type", "application/json")
-	json.NewEncoder(res).Encode(cards)
-}
-
 func ApiCall() {
 	res, err := http.Get(SkcBaseUrl + CardInfoEndpoint)
 	if err != nil {
